181/204 render an array of react elements
Not a large change today, but the change is a result
of doing some reading on JSX outside the book and doing
some experiments.

Tomorrow starts a new chapter, but I may return to this
JSX chapter in the future. An interesting project wouild
be to add JSX to the bold.js file.