188/211 working on script snippets with babel
Specifically, with babel-standalone. which is failing
on a CORS issue when reading directly from a file instead
of a server. This is pretty interesting, and I'm not sure
how far I'm going to take this one. At some point further
into the book I'll be setting up a node server, and perhaps
may revisit this at that time.

Tomorrow, it's worth taking a few minutes to dig a bit
deeper though.