185/206 some annotations in provided source
From yesterday, some investigation into babel configuration
convinced me it's not something I want to deal with at the
moment. There is a lot to it, and won't be learned in a
session or two.

This chapter will apparently be best worked on via the
companion source code to the book. This commit provides
a page number reference in the source to the relevant
discussion in the book.

Tomorrow: continue reading and annotating the source code.

Also: small thunderstorm this morning. It's 75 degrees inside
at 3:30 am.